data frame basic structure
